+++ agent-control/src/lib.rs
@@ -10,6 +10,7 @@ pub mod cli;
 pub mod command;
 pub mod config_migrate;
 pub mod data_store;
+pub mod environment;
 pub mod event;
 pub mod http;
 pub mod instrumentation;